brain/infrastructure/service/brain_telegram_api/requirements.txt
httpx==0.25.0
pyyaml==6.0.1
python-dotenv==1.0.0
watchdog==3.0.0
